Second-order execution now has MIC assigned correctly
To make this work, the service must launch itself with a cloned token which has the mandatory integrity policy flipped off.  win32k initializes certain structures on process creation, and even if it didn't, a process token cannot be altered while it's in use.  Thus, we create a cloned token, and a new process to go with it, and we place that process right on the user's desktop.